requirements.txt
langchain
python-dotenv
langchain_community
streamlit
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
faiss-cpu
langchain_chroma
duckdb
pandas
langchain-groq
langchain_huggingface
duckduckgo-search
validators==0.28.1
youtube_transcript_api
pytube
unstructured
numexpr
pygooglenews
